+++ src/index.ts
@@ -19,3 +19,4 @@ export * from "./usermessages";
 export * from "./sharecode";
 export * from "./crosshair";
 export { parseBinaryKeyValues } from "./keyvalues";
+export { extractPublicEncryptionKey } from "./replaykeys";